feat(board): add rough dedicated utils
